Adds query by timestamp range
This adds support to lookup spans by timestamp range, using the native
date type which in VoltDB is conveniently microsecond capable.

This also polishes a few things, notably that the span json is stored as
a string which makes it possible to use with SQL operators. Also, this
ignores duplicate span upload requests rather than failing.