Adds link aggregation procedure (#9)
This aggregates a link per trace then stores it into the DependencyLink
table. This is not called by anything at the moment, but it can be used
for on-demand dependency link aggregation later.

Notably, this executes Zipkin library code inside VoltDB. You'll notice
we use classes in the zipkin2 package.