Adds trace completion procedure (#10)
This introduces two tables `PendingTrace` and `CompleteTrace`.
When a span is stored, it signals `PendingTrace` via UPSERT that there
was an update. When `CompletePendingTrace` is invoked, it scans through
the `PendingTrace` table and removes entries which correspond to
"completed" traces.

The completion heuristic is a combination of passing a minimum time
threshold since last update (ex 5s), and some data quality checks.
These checks currently are:
  * all spans have timestamp and duration
  * all span ids have a corresponding parent
  * one and only one root span exists

If any of the above fail, the trace will be skipped until a max timeout
(ex 30s). If either condition is met, the trace will be upserted into
the `CompleteTrace` for further processing, such as dependency linking
or export to another system. If a complete trace has late updates, the
upsert's clearing of `CompleteTrace.process_ts` acts as a signal to
re-process.